feat: Add tests

+++ stats_svc/stats-dto/src/main/java/ru/practicum/stats_dto/ViewStatsDto.java
@@ -9,5 +9,5 @@ public class ViewStatsDto {
 
     String app;
     String uri;
-    int hits;
+    Long hits;
 }
